Separating solvers by implementation.
