+++ gillespy2/solvers/cpp/c_base/SimulationTemplate.cpp
