+++ gillespy2/solvers/cpp/c_base/model.h
