+++ gillespy2/solvers/cpp/c_base/ssa.h
